Separated FreeDrive and Navigation to separate concerns.
