+++ app/src/main/java/illyan/jay/ui/navigation/Navigation.kt
@@ -18,16 +18,9 @@
 
 package illyan.jay.ui.navigation
 
-import android.Manifest
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.Text
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
@@ -36,32 +29,20 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.google.accompanist.permissions.ExperimentalPermissionsApi
-import com.google.accompanist.permissions.PermissionState
-import com.google.accompanist.permissions.PermissionStatus
-import com.google.accompanist.permissions.rememberPermissionState
 import com.mapbox.geojson.Point
 import com.mapbox.maps.CameraOptions
 import com.mapbox.maps.EdgeInsets
 import com.mapbox.maps.plugin.animation.flyTo
-import com.mapbox.maps.plugin.locationcomponent.location
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.navigation.EmptyDestinationsNavigator
-import illyan.jay.R
-import illyan.jay.ui.home.RoundedCornerRadius
 import illyan.jay.ui.home.isSearching
 import illyan.jay.ui.home.mapView
 import illyan.jay.ui.home.sheetState
-import illyan.jay.ui.map.turnOnWithDefaultPuck
 import illyan.jay.ui.menu.BackPressHandler
 import illyan.jay.ui.menu.MenuNavGraph
 import illyan.jay.ui.navigation.model.Place
@@ -136,80 +117,4 @@ fun NavigationScreen(
                 "isAnimationRunning ${sheetState.isAnimationRunning}\n" +
                 "Screen height $screenHeight"
     )
-}
-
-@OptIn(ExperimentalPermissionsApi::class)
-@MenuNavGraph
-@Destination
-@Composable
-fun FreeDriveScreen() {
-    val locationPermissionState = rememberPermissionState(
-        Manifest.permission.ACCESS_FINE_LOCATION
-    )
-
-    val context = LocalContext.current
-    when (locationPermissionState.status) {
-        is PermissionStatus.Denied -> {
-            LocationPermissionDeniedScreen(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(
-                        top = 8.dp,
-                        bottom = 8.dp + RoundedCornerRadius,
-                        start = 8.dp,
-                        end = 8.dp
-                    ),
-                locationPermissionState = locationPermissionState
-            )
-        }
-
-        is PermissionStatus.Granted -> {
-            // TODO start session, then wait for new data to be shown by Mapbox Navigation SDK
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(300.dp)
-                    .padding(32.dp)
-            ) {
-                mapView.location.turnOnWithDefaultPuck(context)
-                Text(text = "Thanks for enabling us to spy on you :) UWU owo")
-            }
-        }
-    }
-}
-
-
-
-@OptIn(ExperimentalPermissionsApi::class)
-@Composable
-fun LocationPermissionDeniedScreen(
-    modifier: Modifier = Modifier,
-    locationPermissionState: PermissionState,
-) {
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        Text(
-            text = stringResource(R.string.location_permission_denied_title),
-            style = MaterialTheme.typography.titleLarge
-        )
-        Text(
-            text = stringResource(R.string.location_permission_denied_description),
-            style = MaterialTheme.typography.bodyMedium
-        )
-        Column(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Button(
-                onClick = { locationPermissionState.launchPermissionRequest() },
-            ) {
-                Text(
-                    text = stringResource(R.string.request_permission),
-                    color = Color.White
-                )
-            }
-        }
-    }
 }
